update delete user

+++ scripts/local_user/delete_single_user.sh
@@ -1,23 +1,29 @@
 USER_ACC_NAME=$1
 
-#check if the user exists here
-#check if its root
+if [ $EUID -ne 0 ]; then #delete users and user's folder equire root privileges
+    echo 'No root privileges detected!'
+    echo 'Please, run this script as root'
+else
 
-/usr/bin/dscl . -delete "/Users/$USER_ACC_NAME";
+    #check if the user exists here
 
-if [ $? -eq 0 ] ; then
-    echo "account $USER_ACC_NAME - deleted"
-else 
-    echo "Error while deleting $USER_ACC_NAME"    
-fi
-
-if [ -e "/Users/$USER_ACC_NAME" ]; then 
-    
-    rm -rf "/Users/$USER_ACC_NAME";
+    /usr/bin/dscl . -delete "/Users/$USER_ACC_NAME";
 
     if [ $? -eq 0 ] ; then
-        echo "/Users/$USER_ACC_NAME - deleted"
+        echo "account $USER_ACC_NAME - deleted"
     else 
-        echo "Error while deleting /Users/$USER_ACC_NAME"    
+        echo "Error while deleting $USER_ACC_NAME"    
     fi
+
+    if [ -e "/Users/$USER_ACC_NAME" ]; then 
+        
+        rm -rf "/Users/$USER_ACC_NAME";
+
+        if [ $? -eq 0 ] ; then
+            echo "/Users/$USER_ACC_NAME - deleted"
+        else 
+            echo "Error while deleting /Users/$USER_ACC_NAME"    
+        fi
+    fi
+
 fi
